+++ restaurant-platform-api/src/modules/notifications/notifications.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import { NotificationsController } from './notifications.controller';
+import { NotificationsService } from './notifications.service';
+import { WebsocketsModule } from '../../websockets/websockets.module';
+
+@Module({
+  imports: [WebsocketsModule],
+  controllers: [NotificationsController],
+  providers: [NotificationsService],
+  exports: [NotificationsService],
+})
+export class NotificationsModule {}