+++ restaurant-platform-api/src/modules/orders/orders.module.ts
@@ -3,9 +3,10 @@ import { OrdersService } from './orders.service';
 import { OrdersController } from './orders.controller';
 import { OrdersCustomerController } from './orders-customer.controller';
 import { WebsocketsModule } from '../../websockets/websockets.module';
+import { NotificationsModule } from '../notifications/notifications.module';
 
 @Module({
-  imports: [WebsocketsModule],
+  imports: [WebsocketsModule, NotificationsModule],
   controllers: [OrdersController, OrdersCustomerController],
   providers: [OrdersService],
   exports: [OrdersService],